requirements.txt
streamlit 
openai
torch 
torchvision --index-url https://download.pytorch.org/whl/cu118  
pillow
transformers==4.56.1
Pillow
requests
accelerate
timm
tokenizers

